requirements.txt
# Auto Transcript Project - 依赖包列表
# 安装命令: pip install -r requirements.txt

# 深度学习框架（人脸识别必需）
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.5.0

# 计算机视觉和图像处理
opencv-python>=4.5.0
Pillow>=8.0.0

# 数值计算
numpy>=1.20.0

# 音视频处理
moviepy>=1.0.3
pydub>=0.25.1

# 语音识别
SpeechRecognition>=3.8.0

# 翻译服务
googletrans==4.0.0rc1

# 可选：音频格式支持增强
# ffmpeg-python>=0.2.0

# 可选：更好的音频处理
# librosa>=0.8.0

# 开发依赖（可选）
# pytest>=6.0.0
# pytest-cov>=2.0.0
# black>=21.0.0
# flake8>=3.8.0